Copy test_object_transformer for new test system
